refactor: split MaterialRegistry into registry + parsing + matching modules
materials/registry.py had grown to 667 lines mixing four concerns:
catalog storage/loading, DataFrame caching, discovery, and a fuzzy
name-resolution engine (_extract_wavelength_range alone was flagged
at complexity 20, file total at 90).

Split into:
- _catalog_parsing.py: pure YAML/filename parsing helpers
  (_levenshtein, _catalog_dir_from_filename, _extract_wavelength_range,
  the latter split further into its two extraction strategies).
- _material_matching.py: the stateless name-resolution engine
  (prefilter/find/score/match_policy/row_to_path), operating on a
  DataFrame rather than the registry instance.
- registry.py: MaterialRegistry now owns only catalog storage,
  caching, and discovery, delegating resolution to
  _material_matching.resolve_with_row.

Public API (MaterialRegistry.instance/register/register_file/
load_catalog/resolve/list_*) and the _resolve_with_row test seam are
unchanged.
